pyproject.toml
[tool.poetry]
name = "mongfontbuilder"
version = "0.2.0"
description = ""
authors = [
    "Kushim Jiang <kushim_jiang@outlook.com>",
    "Liang Hai <lianghai@gmail.com>",
]
packages = [{ include = "mongfontbuilder", from = "lib" }]

[tool.poetry.dependencies]
python = "^3.10"
fontmake = "^3.7.1"
PyYaml = "^6.0.1"
unicodedata2 = "^15.1.0"
tptq-feacomposer = "^1.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
uharfbuzz = "^0.42.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
